[IMP] product_stamp_configurator_sale_purchase: cst
Now supplier cost is rounded using new decimal.precision, to make it
more flexible.

[FIX] product_stamp_configurator: time format

Time was formatting in user unfriendly manner, when it was 100 or
bigger, as it would show as like `1e+02`, instead of 100. This was
caused by `:.2g` formatting option. Now leaving only `g` option to
remove useless zeros at the end and time is rounded before it is being
used in description.