[FIX] package_configurator: compute
Now fit quantity for every component is visible before it is saved. It
was not updated, because it was missing some dependencies.

Also, made sure to not duplicate circulation items, when components are
are being added/removed on configurator as originally it was naively
creating without making sure old records were cleaned up.