[ADD] package_configurator: printing
Model to allow specifying printing houses. It also allows to specify
maximum printing width and length for a sheet. So if printing dimensions
are lower than sheet itself, then those sheets will be considered
smaller in fit quantity calculations!

Printing pricelist is used to calculate costs for printing on materials.

Added print setup support, so it would be possible to specify wastage of
sheets/boxes when using print service.